feat(page types): improve Custom Types page
