+++ packages/adapter-nuxt/package.json
@@ -58,7 +58,7 @@
 	},
 	"dependencies": {
 		"@prismicio/slice-simulator-core": "^0.2.7",
-		"@prismicio/types-internal": "2.0.0-alpha.9",
+		"@prismicio/types-internal": "2.0.0-beta.1",
 		"@slicemachine/plugin-kit": "0.2.1-dev-page-types.4",
 		"common-tags": "^1.8.2",
 		"fs-extra": "^11.1.0",
