+++ packages/manager/src/managers/customTypes/types.ts
@@ -1,0 +1,3 @@
+import type { CustomType } from "@prismicio/types-internal/lib/customtypes";
+
+export type CustomTypeFormat = NonNullable<CustomType["format"]>;
