+++ packages/slice-machine/components/CustomTypeTable/changesPage.tsx
@@ -51,7 +51,7 @@ export const CustomTypeTable: React.FC<CustomTypeTableProps> = ({
   isOnline,
 }) => {
   return (
-    <Box as={"table"} data-tag="legacy">
+    <Box as={"table"} data-legacy-component>
       <thead>
         <tr className="transparent small">
           <Box as={"th"} sx={{ width: firstColumnWidth, fontWeight: 600 }}>
