--- packages/slice-machine/src/components/Breadcrumb/Breadcrumb.css.ts
@@ -1,9 +1,0 @@
-import { colors, sprinkles } from "@prismicio/editor-ui";
-import { style } from "@vanilla-extract/css";
-
-export const root = style([
-  sprinkles({
-    all: "unset",
-    color: colors.grey11,
-  }),
-]);
