+++ packages/slice-machine/src/components/Breadcrumb/Breadcrumb.tsx
@@ -1,12 +1,10 @@
 import { Text } from "@prismicio/editor-ui";
 import { FC, ReactNode } from "react";
 
-import * as styles from "./Breadcrumb.css";
-
 type BreadcrumbProps = Readonly<{
   children: ReactNode;
 }>;
 
 export const Breadcrumb: FC<BreadcrumbProps> = ({ children }) => {
-  return <Text className={styles.root}>{children}</Text>;
+  return <Text color="grey11">{children}</Text>;
 };
