+++ packages/slice-machine/src/features/customTypes/customTypesConfig.ts
@@ -1,0 +1,16 @@
+export const CUSTOM_TYPES_CONFIG = {
+  custom: {
+    title: "Custom Types",
+    blankSlateImage: "/blank-slate-page-types.png",
+    blankSlateDescription:
+      "Page types are models that your editors will use to create website pages in the Page Builder.",
+    errorDescription: "An error occurred while fetching your Custom Types.",
+  },
+  page: {
+    title: "Page Types",
+    blankSlateImage: "/blank-slate-page-types.png",
+    blankSlateDescription:
+      "Page types are models that your editors will use to create website pages in the Page Builder.",
+    errorDescription: "An error occurred while fetching your Page Types.",
+  },
+};
